FMZ-12 Corrected a typo in the identifier FMZIdentifable to FMZIdentifiable across various classes.
This commit addresses a typographical error in the naming of the FMZIdentifable identifier. It has been changed to FMZIdentifiable to adhere to the correct English spelling, improving the readability and maintainability of the code. The changes have been made in the FMZDatabase class and the Video class where this identifier was used. Further, the interface name FMZIdentifable.java has been renamed to FMZIdentifiable.java.