FMZ-15 Update README for FMZ Video Player v2.0.0
Updated the README to reflect the changes and improvements introduced in version 2.0.0 of the FMZ Video Player. The header was updated to include the version, additional explanatory text was added about the software, and changes to the configuration and how to run the program were detailed. A logging of past versions was also added and the plan for the next version was updated. This helps to keep documentation up-to-date and clear for users navigating the project.