FMZ-21 Refactor package structure to include application name
The package structure of the project was adjusted to better reflect the specific application it represents. Originally, the classes were under a general 'com.filipmikolajzeglen' package, but now they have been moved to a 'com.filipmikolajzeglen.fmzvideoplayer' package. This change was made for better structure and code organization, allowing for easier navigation and understanding of the codebase. Furthermore, it helps to prevent any potential conflicts with other potential packages or applications. Along with the package restructuring, necessary updates were made to imports and references within the classes and files as needed.